+++ examples/block_events.ts
@@ -1,7 +1,7 @@
 import { client } from "polkadot/mod.ts"
 
 const result = await client
-  .finalizedBlock()
+  .block()
   .events()
   .run()
