+++ examples/block_extrinsics.ts
@@ -1,7 +1,7 @@
 import { client } from "polkadot/mod.ts"
 
 const result = await client
-  .finalizedBlock()
+  .block()
   .extrinsics()
   .run()
