+++ examples/dynamic/balance.ts
@@ -1,11 +1,11 @@
 import { alice } from "capi"
 import { client } from "polkadot_dev/mod.ts"
 
-console.log(
-  await client
-    .metadata()
-    .pallet("System")
-    .storage("Account")
-    .entry([alice.publicKey])
-    .run(),
-)
+const result = await client
+  .metadata()
+  .pallet("System")
+  .storage("Account")
+  .entry([alice.publicKey])
+  .run()
+
+console.log(result)
