+++ fluent/EventsRune.ts
@@ -1,4 +1,6 @@
-import { Rune } from "../rune/mod.ts"
+import { Rune, RunicArgs, ValueRune } from "../rune/mod.ts"
+import { Hex } from "../util/mod.ts"
+import { BlockRune } from "./BlockRune.ts"
 import { Chain, ClientRune } from "./ClientRune.ts"
 
 export type Events<RuntimeEvent extends RuntimeEventBase = RuntimeEventBase> = EventRecord<
@@ -35,7 +37,32 @@ export namespace EventPhase {
 export class EventsRune<out U, out C extends Chain = Chain>
   extends Rune<Events<C["runtimeEvent"]>, U>
 {
-  constructor(_prime: EventsRune<U, C>["_prime"], readonly client: ClientRune<U, C>) {
+  constructor(
+    _prime: EventsRune<U, C>["_prime"],
+    readonly client: ClientRune<U, C>,
+    readonly block: BlockRune<U, C>,
+  ) {
     super(_prime)
+  }
+
+  indexOfTx<X>(...[extrinsicHex]: RunicArgs<X, [Hex]>) {
+    const extrinsics = this
+      .block
+      .into(ValueRune)
+      .access("block", "extrinsics")
+    return Rune
+      .tuple([extrinsics, extrinsicHex])
+      .map(([extrinsics, extrinsicHex]) => {
+        const i = extrinsics.indexOf(("0x" + extrinsicHex) as Hex)
+        return i === -1 ? undefined : i
+      })
+  }
+
+  filterTx<X>(...[hex]: RunicArgs<X, [Hex]>) {
+    return Rune
+      .tuple([this, this.indexOfTx(hex)])
+      .map(([events, i]) =>
+        events.filter((event) => event.phase.type === "ApplyExtrinsic" && event.phase.value === i)
+      )
   }
 }
